moved useDebtSnapshotHistoryQuery.tsx to queries/debt
